Filter out more words with hard to remember spelling
Words that sound similar with different spelling are bad for
passphrases: “Was it a plural or a singular? Maybe with an ‘e’ at the
end? Oh damn, could it be the verb at third person plural… I just can’t
remember.”

French spelling is terribly hard, and it’s even harder
for passphrases as often you can’t see a passphrase being typed.

So let's remove words with endings containing silent letters
and many conjugated forms for verbs.